+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 1.1.0
+*	Add all payment methods to the payment mean brand list of no payment method defined.
+*	Added support for the iDEAL payment method.
+*	Added support for the Incasso (Direct Debit) payment method.
+*	Added an payment mean brand list class.
+
 ## 1.0.0
 *	First release.